+++ explorer/explorer.ts
@@ -3,28 +3,40 @@ import { NodeBase } from './models/nodeBase';
 
 import { config } from '../extension';
 import { WorkspaceNode } from './models/workspaceNode';
+import { AtelierAPI } from '../api';
 
 export class ObjectScriptExplorerProvider implements vscode.TreeDataProvider<NodeBase> {
   onDidChange?: vscode.Event<vscode.Uri>;
   private _onDidChangeTreeData: vscode.EventEmitter<NodeBase> = new vscode.EventEmitter<NodeBase>();
   readonly onDidChangeTreeData: vscode.Event<NodeBase> = this._onDidChangeTreeData.event;
-  private _showSystem = false;
-  private _showSystem4Workspace: boolean[] = [];
+  private _showExtra4Workspace: string[] = [];
 
-  constructor() {}
+  constructor() { }
 
-  get showSystem(): boolean {
-    return this._showSystem;
+  async selectNamespace(workspaceFolder: string): Promise<any> {
+    let api = new AtelierAPI(workspaceFolder);
+    return api
+      .serverInfo()
+      .then(data => data.result.content.namespaces)
+      .then(data => data.filter(ns => ns !== api.ns && !this._showExtra4Workspace.includes(ns)))
+      .then(data => data.map(ns => ({ label: ns })))
+      .then(vscode.window.showQuickPick)
+      .then(ns => this.showExtra4Workspace(workspaceFolder, ns.label));
   }
 
-  set showSystem(value) {
-    this._showSystem = value;
-    this._onDidChangeTreeData.fire(null);
+  showExtra4Workspace(workspaceFolder: string, ns: string) {
+    if (!this._showExtra4Workspace.includes(ns)) {
+      this._showExtra4Workspace.push(ns);
+      this._onDidChangeTreeData.fire(null);
+    }
   }
 
-  showSystem4Workspace(workspaceFolder: string, value: boolean) {
-    this._showSystem4Workspace[workspaceFolder] = value;
-    this._onDidChangeTreeData.fire(null);
+  closeExtra4Workspace(workspaceFolder: string, ns: string) {
+    let pos = this._showExtra4Workspace.indexOf(ns);
+    if (pos >= 0) {
+      this._showExtra4Workspace.splice(pos, 1)
+      this._onDidChangeTreeData.fire(null);
+    }
   }
 
   refresh(): void {
@@ -53,10 +65,10 @@ export class ObjectScriptExplorerProvider implements vscode.TreeDataProvider<Nod
         node = new WorkspaceNode(workspaceFolder.name, this._onDidChangeTreeData);
         rootNodes.push(node);
 
-        if (this.showSystem || this._showSystem4Workspace[workspaceFolder.name]) {
-          node = new WorkspaceNode(workspaceFolder.name, this._onDidChangeTreeData, true);
+        this._showExtra4Workspace.forEach(ns => {
+          node = new WorkspaceNode(workspaceFolder.name, this._onDidChangeTreeData, ns);
           rootNodes.push(node);
-        }
+        })
       }
     });
     return rootNodes;